Make the test suite easier to extend
From the readability review:

- Session-scoped wheel/package fixtures now live in conftest files
  instead of traveling via cross-module noqa imports; the fixture
  module is renamed test_packages.py -> package_fixtures.py (it
  contains no tests) and api support code converter.py ->
  converter_support.py (no longer shadows the source module)
- CliTestCaseFactory.__call__ forwards keyword arguments to
  CliTestCase instead of re-declaring every expected_* parameter;
  the never-asserted expected_interactive value is removed
- test_converter.py (1900+ lines) is split along its existing
  section boundaries into test_requires_dist.py,
  test_platform_target.py, test_binary_deps.py, and
  test_dependency_translation.py
- PackageValidator's independent-oracle design is now documented in
  its class docstring, and __call__ is defined next to validate
