+++ test/api/conftest.py
@@ -1,0 +1,19 @@
+#  Copyright 2026 Christopher Barber
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+"""
+Shared fixtures for whl2conda.api tests
+"""
+
+from .converter_support import test_case  # noqa: F401
+from .validator import validate_package  # noqa: F401
