+++ test/conftest.py
@@ -73,3 +73,13 @@ def pytest_collection_modifyitems(config, items):
         for item in items:
             if "slow" in item.keywords:
                 item.add_marker(skip_slow)
+
+
+# Shared session-scoped wheel/package fixtures, made available to the
+# whole suite here so that individual test modules need not import them.
+from .package_fixtures import (  # noqa: F401
+    markers_wheel,
+    setup_wheel,
+    simple_conda_package,
+    simple_wheel,
+)
